调整PID分辨率

+++ include/base/embedded/timer/InputCaptureTimerPll.cpp
@@ -70,7 +70,7 @@ base::InputCaptureTimerPll::InputCaptureTimerPll(base::input_capture_timer::Inpu
 		base::Int64Fraction{1, 100},
 		base::Int64Fraction{1, 1000},
 		0,
-		base::Int64Fraction{1, INT16_MAX},
+		base::Int64Fraction{1, INT32_MAX},
 		static_cast<int64_t>(_timer.CounterPeriod() / 2),
 		-static_cast<int64_t>(_timer.CounterPeriod() / 2),
 	};
